fix: Crm table revert to component level and updated
