+++ frontend/src/community/crm/constants/contactConstants.ts
@@ -1,7 +1,7 @@
 export const CONTACT_SEARCH_DEBOUNCE_DELAY = 500;
 export const DEFAULT_PAGE_SIZE = 10;
 export const DEFAULT_COMPANY_PAGE_SIZE = 100;
-export const ALL_COMPANIES = undefined;
+export const ALL_COMPANIES = "";
 export const CONTACT_NAME_MAX_LENGTH = 255;
 export const CONTACT_EMAIL_MAX_LENGTH = 100;
 export const CONTACT_NUMBER_MAX_LENGTH = 15;
